fix unit tests/linting
